requirements.txt
# Korean-Vietnamese Parallel Corpus Pipeline Dependencies

# Web Crawling
# Korean-Vietnamese Parallel Corpus Pipeline Dependencies

# Web Crawling
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
lxml[html_clean]

# Text Extraction
# Text Extraction
trafilatura>=1.6.0

# Korean NLP
# Korean NLP
kss>=5.2.0  # Korean sentence splitter
kiwipiepy>=0.18.0  # Korean tokenizer

# Vietnamese NLP
# Vietnamese NLP
underthesea>=6.8.0  # Vietnamese NLP toolkit

# Embeddings & ML
# Embeddings & ML
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Language Detection
# Language Detection
fasttext>=0.9.2
# Note: Download fastText LID model separately:
# wget https://dl.fbaipublicfiles.com/fasttext/supervised-models/lid.176.bin

# Utilities
# Utilities
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0

# Optional: GPU support
# Optional: GPU support
# If you have CUDA, install:
# torch with CUDA: pip install torch --index-url https://download.pytorch.org/whl/cu118